[improve](AllPages) Improving performance and change file structure
