+++ src/components/AboutComponents/AboutNavbar/AboutNavbar.tsx
@@ -6,7 +6,6 @@ export default function AboutNavbar() {
   useEffect(() => {
     function handleScroll() {
       setIsScrolled(window.scrollY > 0);
-      console.log(window.scrollY);
     }
 
     window.addEventListener("scroll", handleScroll);
