+++ src/components/DashboardComponents/ExportDashboard/options/FilterOptions.tsx
@@ -1,7 +1,7 @@
 import { Accordion, AccordionTab } from "primereact/accordion";
-import { filterOptions, useExportHook } from "../../../hooks/useExportHook";
-import { Campus, campus_to_campuscode, ReportType, reporttype_to_string, type ReportData } from "../../../types/variables";
-import { capitalize } from "../../../utils/other";
+import { filterOptions, useExportHook } from "../../../../hooks/useExportHook";
+import { Campus, campus_to_campuscode, ReportType, reporttype_to_string, type ReportData } from "../../../../types/variables";
+import { capitalize } from "../../../../utils/other";
 
 export default function FilterOptions() {
   const { filter, setFilter } = useExportHook();
