+++ src/components/DashboardComponents/TableDashboard/Mobile/ReportMobileTable.tsx
@@ -1,6 +1,6 @@
-import { useReportDataHook } from "../../../hooks/shared/useReportData";
-import { useReportFilterHook, useReportPaginationHook } from "../../../hooks/useReportHook";
-import LoadingAnimation from "../../GlobalComponents/Loading/LoadingAnimation";
+import { useReportDataHook } from "../../../../hooks/shared/useReportData";
+import { useReportFilterHook, useReportPaginationHook } from "../../../../hooks/useReportHook";
+import LoadingAnimation from "../../../GlobalComponents/Loading/LoadingAnimation";
 import ReportMobileCard from "./ReportMobileCard";
 
 export default function ReportMobileTable() {
