--- src/components/DashboardComponents/TableDashboard/ReportTableWrapper.tsx
@@ -1,19 +1,0 @@
-import React, { Suspense } from "react";
-import { ReportHookEffect } from "../../../hooks/useReportHook";
-import ReportTableToolbar from "./ReportTableToolbar";
-import LoadingAnimation from "../../GlobalComponents/Loading/LoadingAnimation";
-
-const ReportListComponent = React.lazy(() => import("./ReportTableComponent"));
-
-export default function ReportTableWrapper() {
-
-    return <>
-        <ReportHookEffect />
-        <div className="space-y-4">
-            <ReportTableToolbar />
-            <Suspense fallback={<LoadingAnimation />}>
-                <ReportListComponent />
-            </Suspense>
-        </div>
-    </>
-}
