--- src/pages/MainPages/_FormReportPage.tsx
@@ -1,10 +1,0 @@
-import ReportFormComponent from "../../components/ReportFormComponents/FormReport/ReportFormComponent";
-
-export default function ReportForm() {
-
-  return (
-    <>
-      <ReportFormComponent />
-    </>
-  )
-}
