+++ src/pages/MainPages/_wrapper.tsx
@@ -1,20 +1,36 @@
 import { useEffect, useState } from 'react';
 import MainPage from './_index';
 import WelcomeComponent from '../../components/ReportFormComponents/WelcomeOverlay/WelcomeComponent';
+import SelectCampusOverlay from "../../components/ReportFormComponents/SelectCampus/SelectCampusOverlay";
+import { useCampusDataHook } from "../../hooks/shared/useCampusData";
 
 export default function WrapperPage() {
-    const [showMain, setShowMain] = useState(true);
+    const [showWelcome, setShowWelcome] = useState(false);
+    const { selectedCampus } = useCampusDataHook();
 
     useEffect(() => {
         if (!sessionStorage.getItem("no-welcome-page")) {
             sessionStorage.setItem("no-welcome-page", "true");
-            setShowMain(false);
+            setShowWelcome(true);
         }
     }, []);
 
     return (
         <>
-            {showMain ? <MainPage /> : <WelcomeComponent onFinish={() => setShowMain(true)} />}
+            {(() => {
+                // If the user should see the welcome animation
+                if(showWelcome) {
+                    return <WelcomeComponent onFinish={() => setShowWelcome(true)} />;
+                }
+                
+                // If user hasn't select any campus
+                if(!selectedCampus) {
+                    return <SelectCampusOverlay />;
+                }
+
+                // If there's no welcome page, and the user has select a campus
+                return <MainPage />;
+            })()}
         </>
     );
 }
