--- src/pages/dashboard/DashboardPage/_TablePage.tsx
@@ -1,9 +1,0 @@
-import ReportTableWrapper from "../../../components/DashboardComponents/TableDashboard/ReportTableWrapper";
-
-export default function TablePage() {
-    return (
-        <>
-            <ReportTableWrapper />
-        </>
-    );
-}
