--- src/pages/dashboard/DashboardPage/_UsersPage.tsx
@@ -1,14 +1,0 @@
-import React, { Suspense } from "react"
-import LoadingAnimation from "../../../components/GlobalComponents/Loading/LoadingAnimation"
-
-const UsersComponent = React.lazy(() => import("../../../components/DashboardComponents/UsersDashboard/UsersTabComponent"))
-
-export default function UsersPage() {
-    return (
-        <>
-            <Suspense fallback={<LoadingAnimation />}>
-                <UsersComponent />
-            </Suspense>
-        </>
-    )
-}
