--- src/pages/dashboard/DashboardPage/_WelcomePage.tsx
@@ -1,7 +1,0 @@
-import WelcomeComponent from "../../../components/DashboardComponents/WelcomeDashboard/WelcomeComponent";
-
-export default function WelcomePage() {
-    return (
-        <WelcomeComponent />
-    )
-}
