+++ src/pages/dashboard/index.astro
@@ -1,9 +1,47 @@
 ---
-// admin/index.astro
+// Pages
+import WelcomePage from "../../components/DashboardComponents/WelcomeDashboard/WelcomePage";
+import ReportTablePage from "../../components/DashboardComponents/TableDashboard/ReportTablePage";
+import StatisticsPage from "../../components/DashboardComponents/StatisticsDashboard/StatisticsPage";
+import ExportPage from "../../components/DashboardComponents/ExportDashboard/ExportPage";
+import UsersPage from "../../components/DashboardComponents/UsersDashboard/UsersPage";
+import SettingPage from "../../components/DashboardComponents/SettingDashboard/SettingPage";
+
+// Others
 import Layout from "../../layouts/Layout.astro";
-import DashboardPage from "./DashboardPage/_index";
+import SidebarDashboard from "../../components/DashboardComponents/SidebarAdmin/SidebarAdmin";
+import UseUserDataHookEffect from "../../hooks/shared/useUserData";
+import UseReportConfigHookEffect from "../../hooks/useReportConfig";
+import SidebarToggle from "./DashboardPage/_SideBarToggle";
 ---
 
 <Layout pageType="dashboard">
-  <DashboardPage client:only />
+  <UseUserDataHookEffect client:load adminOnly />
+  <UseReportConfigHookEffect client:load useAll />
+  
+  <div class="flex flex-col h-screen p-4 border 2">
+      {/* Top Navbar */}
+      <div class="bg-white rounded-2xl shadow shadow-gray-500 z-10 flex justify-between items-center px-4 py-1 sticky top-0">
+          <div>
+              <SidebarToggle client:load />
+              <img src="/img/logoSekolah.png" class="w-10 h-auto ring-2 ring-white rounded-full" alt="" />
+          </div>
+      </div>
+
+      {/* Main Area: Sidebar + Scrollable Content */}
+      <div class="flex gap-4 py-4 px-1 h-full overflow-y-auto">
+          {/* Sidebar */}
+          <SidebarDashboard client:load />
+
+          {/* Scrollable Content */}
+          <div class="flex-1 overflow-y-auto">
+              <WelcomePage client:load />
+              <ReportTablePage client:load />
+              <StatisticsPage client:load />
+              <ExportPage client:load />
+              <UsersPage client:load />
+              <SettingPage client:load />
+          </div>
+      </div>
+  </div>
 </Layout>
